pyproject.toml
[project]
name = "dataforge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.15",
    "alembic>=1.15.2",
    "apscheduler>=3.11.0",
    "chainlit>=2.5.5",
    "chromadb>=0.6.3",
    "faker>=37.1.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "langchain>=0.3.22",
    "langchain-community>=0.3.20",
    "langchain-mcp-adapters>=0.0.7",
    "langchain-ollama>=0.3.2",
    "langchain-openai>=0.3.12",
    "langfuse>=2.60.2",
    "langgraph>=0.3.21",
    "langgraph-cli[inmem]>=0.1.83",
    "langgraph-prebuilt>=0.1.8",
    "loguru>=0.7.3",
    "markdown>=3.7",
    "mcp[cli]>=1.6.0",
    "multidict>=6.3.0",
    "openai>=1.70.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "ruamel-yaml>=0.18.10",
    "ruff>=0.11.8",
    "sqlalchemy>=2.0.40",
    "sqlglot[rs]>=26.12.1",
    "streamlit>=1.44.1",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
name = "aliyun"
default = true
